Moved schemas into a top level directory.
It seems strange to have applications using this library to need to
reach into the src directory, so I added a schemas directory on the top
level and moved the fbs files there.

Additionally, the CMakeLists file now has a variable containing the
location of schemas that can be used by the project including pindrop.

Change-Id: I384d57ba08f0fb913db2f5b6cc32f0d06d7cfdc2