Fixed the sorting comparators.
Previously, the comparators to sort priorities of playing sounds
operated like strcmp. That is, they would return 1, 0, or -1 depending
on if the priority was less than, equal to, or greater than the sound
being compared to. `std::sort` actually expects a function that returns
a less-than comparison. Miraculously, the incorrect code usually worked,
but would occasionally crash when sorting elements of the list.

This change fixes the comparators. Additionally, more robust unit tests
have been added to verify that they sort correctly. The old code did not
pass these unit test, the new code does. Tested on Linux.

Bug: b/19005164
Change-Id: I61cc71817f5d5c7f4e40b160d3d023162c606c26