The list of playing sounds now an intrusive list.
Previously, the list of playing sounds was a vector that was sorted by
priority every frame. Now the list of playing sounds is an intrusive
list which has elements added to it in the proper order so sorting is no
longer necessary.

The intrusive list implementation (found in intrusive_list.h) was
adapted from LiquidFun. References to Box2D were removed and it was
formatted with ClangFormat to match the code style of Pindrop.

Tested on linux with Pie Noon. Added unit tests which all pass.
Bug: 18276830

Change-Id: Ic07489aed8fc3ef2f336fe59fb3dbba0e445bc16