Replaced ref counter on buses with intrusive list.
Previously, buses kept track of how many sounds were playing on that bus
with a ref counter. That has been switched to an IntrusiveList. To make
this work, PlayingSounds now have two IntrusiveListNodes, one for the
priority list and one for the bus list.

Additionally, I noticed a bug while adding this: when playing a stream
it would not be tracked by the bus and so if it was supposed to duck
other buses nothing would happen. A single PlayingSound was added to the
AudioEngineInternalState to track the currently playing stream.

Also removed some obsolete functions.

Bug: 18641604
Tested on Linux, all tests pass. Also works in Pie Noon.

Change-Id: I75f4dbee55845e047889249bba5cf6fb54bb7666