Added distance based gain dampening.
SoundCollections now have two modes, nonpositional and positional.
Nonpositional sounds simply play at the volume specified. Positional
sounds however may have their gain attenuated based on the sound's
distance from the listener.  SoundCollectionDefs now have a variety of
parameters that control how far away a sound may be heard from and the
shape of the drop off curve.

Additionally you may specify a set of listeners, which are simply points
in space. The closest listener to the sound is the one used for distance
calculations.

Added a Sort function for Intrusive lists so they can sort the
prioritized channels. Added unit tests for the sort functionality.

Also fixed a bug in the unit tests on Windows.

Unit tests have been added for the new functionality.

Tested on Linux with Pie Noon. Ensured that it compiled in windows.

Change-Id: Icd64c2f8762be2c565d5ecc2b4e909b2d90ac50a