Renamed audio_engine.h to pindrop.h.
Initially the file was named audio engine because I didn't have a name
for the library yet, and that is all the file held. It now contains two
other structures and a handful of typedefs and other things, so it seems
like it ought to be renamed.

Built and testesd on linux.

Change-Id: If89c9ec268a5601b1468c767a0fa64c10d3409bf