Refactored sound code.
Sounds had an unnecessary function to set gain (which Channels are
supposed to handle) so I removed it. also moved around some of the code
that handles playing sounds.

The idea is that audio_engine.cpp should not have to care about the
underlying audio backend it depends on or the implementation details of
the channels or sounds so that in the future hooking up to other audio
backends will be easier.

Tested on Linux.

Change-Id: Ie763b6b8b81833c2f09677fced54eb04abac0d16