Don't invalidate sounds that have stopped playing.
Sounds that have stopped should not be marked invalid. Making them invalid
means that querying their Playing status won't work.

Due to the way SDL Mixer handles it's single audio stream, Pindrop needs
to track the channel that's assigned to it so that it can manually clean
up channel data.

Additionally made some minor fixes to ensure it works on Windows.

Tested on Windows and Linux with Pie Noon.

Change-Id: Ic187a3dd858c3da97b601ff20933146a1945574f