Added panning.
Listeners are now backed by a transformation matrix instead of a single
vector representing location. This allows them to have both a location
and an orientation which is needed for panning. The function SetMatrix
has been added to manipulate this matrix directly, but a helper
SetOrientation function has been added as well if you just want to
supply a location, direction, and up vector instead.

There was also a bunch of refactoring of the audio engine internals, and
unit tests were added for panning.

The sample project now support panning as well.

Tested on Linux and Windows. All unit tests pass.

Change-Id: I681998758a4c2e73fe71f2ca55316bb39adfc39c