Do not attempt to play sounds when paused.
When the audio engine is paused, the audio engine will no longer attempt
to play sounds. Instead, the sound will be placed in a virtual channel
if one is available, and the code that that manages real channels will
be skipped until playback resumes.

Tested with Pie Noon on Linux and Android.

Change-Id: I694c4515b85eb071e85876627130315b57bb6d1c