Fixed nonpositional audio panning.
When using nonpositional audio the panning was set based on an
uninitialized vector. This was fine in some debug builds because the
vector was initialized to 0, but would cause odd panning effects in
other cases (like the panning rapidly switching between left and right
speakers).

Tested on Linux and Android.

Change-Id: Ib4e74312e50bdb24173cd1e0866e71db5ef3bf86