Added the option to specify a user specified gain.
The user can now access buses, and has the ability so apply a gain to
them, as well as to individual channels.

What was previously known as a Bus has been renamed BusInternalState,
and the Bus class is now a PIMPL for the BusInternalState (the same
pattern that's used for the Channel and Listener classes).

Updated the Demo project to use adjust the Bus's gain based on X
location of the mouse as an example of how it can be used (and to serve
as a test case to make sure it worked).

Tested on Linux and Android.
Bug: 22456141

Change-Id: I073ce4a288cb477600db7b4675704719480e7f0d