Fixed a bug in panning audio positioned at (0,0,0)
In cases where positional audio was being played at (0, 0, 0) the
panning code would end up with panning values of NaN.

Addititionally, the matrix math on sound locations was being done
incorrectly, causing audio to be translated in strange ways. This was
fixed by multiplying by the inverse of the matrix representing the
listener rather than the matrix itself. This was changed to store the
inverse matrix becuase it's queried more often then it's set, so doing
the inversion all the time would be wasteful.

Tested on Linux and OSX
Bug: 22874159

Change-Id: I17f2d34e4ea68ec2397831737c54122614fcb28f