Fixed crash bug.
When organizing the real channels among the prioritized list of sounds,
there was a possible edge case where a sound that was fading out would
not be properly halted, which would put it in an inconsistent state and
violate assumptions elsewhere in the code resulting in an assert.

Tested on Linux.

Change-Id: I1a1683676178f79ccc3f8024ecab2084421601fe