Fixed an error in state handling.
There is a case where a low priority paused sound can become
virtualized, and when it becomes devirtualized later the code to play
the sound asserted because it does not see it as playing.

The correct behavior is to begin playing the sound if the channel was
previously playing and to start the sound up but paused if it was
virtualized when it was paused.

Also the assert logic not really necessary since I need to be able to
play a sound that's paused so that I can actually pause it.

Tested on Linux and Android.
Bug: 25340448
Change-Id: I04f5f2b8d14bf78196016e14b7f8714b66966847