Add fplbase to list of pindrop dependencies.
At the moment SDL inside the fplbase project is used to build a
static variant of the library which is required for SDL_mixer so we
need to include fplbase.  In future we should remove this dependency
on fplbase as Pindrop does not need it.

Tested:
Verified pindrop builds in the distro source layout.
Bug: 25780586

Change-Id: I5953011a6d6a21a21b5e336329d2fa75cd697b55