Ensure old stream halts when new stream plays.
There was a bug where under certain conditions streaming audio would
continue to play even though the engine had paused.

When a stream begins playing, a virtual channel tracks it. Then a new
stream plays, halting the previous stream. However, the original virtual
channel still points at it. Now there are two channels pointing at the
stream channel which screws things up.

This change makes it so that only one channel can ever be considered to
be the one that is playing streaming audio, which ensures that the
channels are tracked properly.

Tested on Android on Pie Noon.
Bug: 25875017

Change-Id: I48e08715857c9acbf5a2806789c7be681cb4d858